Move avifCalloc() to internal.h and use libc calloc()
Addressed:
- Move avifCalloc() declaration from the public avif.h to internal.h
  and drop AVIF_API; avifAlloc()/avifFree() were exported in the public
  interface unintentionally and the new helper is for internal use only.
- Update the declaration comment to describe behavior (incl. the
  overflow case) rather than the rationale for the helper.
- Simplify the body to a single call to libc calloc(), which already
  handles count*size overflow per C11. Drop the now-unused <stdint.h>
  and <string.h> includes.
- codec_svt.c: pass uvSize as the element count, sizeof(uint8_t) as
  the element size, matching the count-then-size convention.
- gainmap.c: drop the redundant (size_t) cast on numBuckets.

+++ src/gainmap.c
@@ -393,7 +393,7 @@ avifResult avifFindMinMaxWithoutOutliers(const float * gainMapF, size_t numPixel
 
     const int maxNumBuckets = 10000;
     const int numBuckets = AVIF_MIN((int)ceilf((max - min) / bucketSize), maxNumBuckets);
-    int * histogram = avifCalloc((size_t)numBuckets, sizeof(int));
+    int * histogram = avifCalloc(numBuckets, sizeof(int));
     if (histogram == NULL) {
         return AVIF_RESULT_OUT_OF_MEMORY;
     }
